Move UserEmailAddress to Shoebox
